refactor: reduce ManipulationHandle3D.java from 664 to 487 lines
Extract logical groups into package-private delegates:
- Color4fInterruptibleAnimation: top-level animation class
- DoubleInterruptibleAnimation: top-level animation class
- HandleGeometryHelper: geometry, scale, bounding box, camera opacity,
  and pick-criterion logic

Public API unchanged. All 357 story-api tests pass.

Closes #623
